[tests] Remove spurious error log in p2p_segwit.py
Since 39c18a01475d8b90346c71b989eca5477f3685f4, when wait_until() fails,
an error message is logged to the test framework log. This means that if
wait_until() is called inside a try-except with the expectation that it
will fail, a spurious error message is logged.

wait_until() shouldn't be called with the expectation of failure. Fix
that in p2p_segwit.py.