[addrman] Fix new table bucketing during unserialization
An addrman entry can appear in up to 8 new table buckets. We store this
entry->bucket indexing during shutdown so that on restart we can restore
the entries to their correct buckets.

Commit 5ca91145ec1f746e459a6c0117bc5038108115aa broke the
deserialization code so that each entry could only be put in up to one
new bucket. Fix that.